requirements.txt
# ============================================================================
# HybridRAG v3 — Python Dependencies
# ============================================================================
# Python Version: 3.10+ (tested on 3.10.1 and 3.11.9)
#
# INSTALL ORDER MATTERS:
#   Step 1: pip install torch==2.5.1+cpu --index-url https://download.pytorch.org/whl/cpu
#   Step 2: pip install -r requirements.txt
#
# PyTorch must be installed first with the CPU-only index URL.
# If you skip Step 1, pip will download the full GPU version (~2GB).
#
# Total download: ~800MB (PyTorch ~200MB, models ~80MB on first run)
# ============================================================================

# --- Core ML / Embedding ---
# These power the semantic search pipeline
sentence-transformers==2.7.0       # Text → 384-dim vector embeddings (all-MiniLM-L6-v2)
torch==2.10                        # PyTorch CPU-only (required by sentence-transformers)
transformers==4.57.6               # Hugging Face model loading
tokenizers==0.22.2                 # Fast tokenization (used by transformers)
safetensors==0.7.0                 # Safe model weight loading (replaces pickle)
huggingface_hub==0.36.1            # Model download from Hugging Face
numpy==1.26.4                      # Numerical arrays (embeddings, dot products)
scikit-learn==1.8.0                # Cosine similarity, cross-encoder utilities
scipy==1.17.0                      # Sparse matrix ops (used by scikit-learn)

# --- Document Parsers ---
# Each one handles a specific file format
pdfplumber==0.11.9                 # PDF text extraction (primary PDF parser)
pdfminer.six==20251230             # PDF internals (used by pdfplumber)
pypdf==6.6.2                       # PDF metadata and page counting
pypdfium2==5.3.0                   # Fast PDF rendering (backup parser)
python-docx==1.2.0                 # Word .docx parsing
python-pptx==1.0.2                 # PowerPoint .pptx parsing
openpyxl==3.1.5                    # Excel .xlsx reading
xlsxwriter==3.2.9                  # Excel .xlsx writing (reports)
lxml==6.0.2                        # XML/HTML parsing (used by docx/pptx)
et_xmlfile==2.0.0                  # XML handling (used by openpyxl)
pillow==12.1.0                     # Image processing (OCR prep, thumbnails)

# --- OCR (optional — requires Tesseract installed separately) ---
pytesseract==0.3.13                # Python wrapper for Tesseract OCR engine
pdf2image==1.17.0                  # Convert PDF pages to images for OCR

# --- HTTP / API ---
httpx==0.28.1                      # HTTP client for Ollama + online API calls
h11==0.16.0                        # HTTP/1.1 protocol (used by httpx)
httpcore==1.0.9                    # HTTP transport (used by httpx)
anyio==4.12.1                      # Async I/O (used by httpx)
requests==2.32.5                   # HTTP client (used by huggingface_hub)
urllib3==2.6.3                     # HTTP utilities (used by requests)
certifi==2026.1.4                  # SSL certificates
idna==3.11                         # International domain names
chardet==5.2.0                     # Character encoding detection
charset-normalizer==3.4.4          # Character encoding (used by requests)

# --- Configuration / Serialization ---
PyYAML==6.0.2                      # YAML config file loading
typing_extensions==4.15.0          # Extended type hints (Python 3.10+ compat)

# --- Logging / CLI / Display ---
structlog==24.4.0                  # Structured logging (JSON-safe, auditable)
rich==13.9.4                       # Pretty terminal output (tables, progress)
colorama==0.4.6                    # Windows terminal color support
tqdm==4.67.3                       # Progress bars
Pygments==2.19.2                   # Syntax highlighting (used by rich)
markdown-it-py==4.0.0              # Markdown rendering (used by rich)
mdurl==0.1.2                       # URL utilities (used by markdown-it-py)

# --- Security ---
cryptography==46.0.4               # AES-256 encryption (future: at-rest encryption)
cffi==2.0.0                        # C bindings (used by cryptography)
pycparser==3.0                     # C parser (used by cffi)

# --- Math / Graph (transitive dependencies) ---
# These are pulled in by torch/transformers but pinned for reproducibility
sympy==1.13.3                      # Symbolic math
mpmath==1.3.0                      # Arbitrary precision math
networkx==3.6.1                    # Graph algorithms
joblib==1.5.3                      # Parallel processing (used by scikit-learn)
threadpoolctl==3.6.0               # Thread pool management
filelock==3.20.3                   # File locking (model downloads)
fsspec==2026.1.0                   # Filesystem abstraction
packaging==26.0                    # Version parsing
Jinja2==3.1.6                      # Templating (used by transformers)
MarkupSafe==3.0.3                  # HTML escaping (used by Jinja2)
regex==2026.1.15                   # Enhanced regex (used by tokenizers)
xxhash==3.5.0                      # Fast hashing (used by huggingface_hub)
tiktoken==0.8.0                    # Token counting (cost estimation)
